Rename Parser to CoreParser
